pyproject.toml
[tool.poetry]
name = "onebudget-api"
version = "0.1.0"
description = ""
authors = ["michaeldcanady <michaeldcanady+github@gmail.com>"]
readme = "README.md"
packages = [{include = "onebudget_api"}]

[tool.poetry.dependencies]
python = "^3.11"
anyio = "3.6.2"
autopep8 = "2.0.2"
bcrypt = "4.0.1"
bleach = "6.0.0"
cffi = "1.15.1"
click = "8.1.3"
colorama = "0.4.6"
cryptography = "39.0.2"
fastapi = "0.94.0"
greenlet = "2.0.2"
h11 = "0.14.0"
idna = "3.4"
psycopg2-binary = "2.9.5"
pycodestyle = "2.10.0"
pycparser = "2.21"
pydantic = "1.10.6"
pyjwt = "2.6.0"
pyotp = "2.8.0"
python-dotenv = "1.0.0"
python-multipart = "0.0.6"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.6"
starlette = "0.26.0.post1"
typing-extensions = "4.5.0"
uvicorn = "0.21.0"
webencodings = "0.5.1"


[tool.poetry.group.tests.dependencies]
pytest = "^7.3.0"
requests = "^2.28.2"


[tool.poetry.group.linting.dependencies]
pylint = "^2.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
